requirements.txt
# LangChain
langchain-classic>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0

# Vector Store
chromadb>=0.5.0

# Embeddings
sentence-transformers>=2.7.0

# Web Scraping
beautifulsoup4>=4.12.0

# LLM - Google Gemini (best free tier: 15 RPM, 1M TPM)
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0
